Creating new files and removing un-needed ones
Cleaned up the branch to better match the agreed upon structure
