--- src/main/kotlin/game/manager/nomic/api/config/NomicConfigProperties.kt
@@ -1,12 +1,0 @@
-package game.manager.nomic.api.config
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "nomic")
-data class NomicConfigProperties(
-    var dbEndpoint: String = "",
-    var dbUsername: String = "",
-    var dbPassword: String = "",
-    var dbPort: Int = -1,
-    var dbName: String = ""
-)
