+++ src/main/kotlin/nomic/api/Games.kt
@@ -1,0 +1,4 @@
+package nomic.api
+
+class Games {
+}
