+++ src/main/kotlin/nomic/data/repositories/games/GameRepository.kt
@@ -1,0 +1,4 @@
+package nomic.data.repositories.games
+
+class GameRepository {
+}
