+++ src/main/kotlin/nomic/data/repositories/games/IGameRepository.kt
@@ -1,0 +1,4 @@
+package nomic.data.repositories.games
+
+interface IGameRepository {
+}
