+++ src/main/kotlin/nomic/domain/games/GameDomain.kt
@@ -1,0 +1,4 @@
+package nomic.domain.games
+
+class GameDomain {
+}
