+++ src/main/kotlin/nomic/domain/games/IGameDomain.kt
@@ -1,0 +1,4 @@
+package nomic.domain.games
+
+interface IGameDomain {
+}
